Refactor agent salesperson message trigger logic into helper function
